test: reorganize framework and output coverage
- split rejection tests from happy-path behavior
- move framework and language tests into grouped folders
- add focused Vue parser coverage
- extract shared stream capture helper
- cover output formatter formats separately
