+++ test/frameworks/svelte.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from "vitest";
 
-describe("Svelte traversal", () => {
+describe("Svelte file collection", () => {
   it.todo("collects dependencies declared in Svelte <script> blocks");
 });
